refactor : input/output main methodë로 코드 수정 및 클래스 리네이밍
