--- src/main/java/lotto/WebUILottoApplication.java
@@ -1,12 +1,0 @@
-package lotto;
-
-import lotto.controller.Lotto;
-
-import static spark.Spark.get;
-
-public class WebUILottoApplication {
-
-    public static void main(String[] args) {
-        Lotto.lotto();
-    }
-}
